requirements.txt
autopep8==1.5.7
psycopg2==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
toml==0.10.2
